feat: moved components into app
