+++ src/app/components/Card/Card.module.scss
