+++ src/app/components/Card/Card.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
-import selectionReducer from '../../features/selectionSlice';
+import selectionReducer from '../../../features/selectionSlice';
 import Card from './Card';
 import userEvent from '@testing-library/user-event';
