+++ src/app/components/CardList/CardList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../Card/Card';
+import Card from '../../app/components/Card/Card';
 import CardListSkeleton from '../CardListSkeleton/CardListSkeleton';
 import s from './CardList.module.scss';
