+++ src/app/components/CardListSkeleton/CardListSkeleton.module.scss
