+++ src/app/components/CardSkeleton/CardSkeleton.module.scss
