+++ src/app/components/ErrorBoundary/ErrorBoundary.module.scss
