+++ src/app/components/ItemDetail/ItemDetail.module.scss
