+++ src/app/components/MainBlock/MainBlock.module.scss
