+++ src/app/components/MainBlock/MainBlock.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect, useCallback, useMemo } from 'react';
 import { Outlet, useMatch, useNavigate, useLocation } from 'react-router-dom';
-import useLocalStorage from '../../hooks/useLocalStorage';
+import useLocalStorage from '../../../hooks/useLocalStorage';
 import CardList from '../CardList/CardList';
 import Header from '../Header/Header';
-import Pagination from '../Pagination/Pagination';
+import Pagination from '../../../../../Pagination/Pagination';
 import SelectionBlock from '../SelectionBlock/SelectionBlock';
 import s from './MainBlock.module.scss';
 
@@ -12,11 +12,11 @@ import {
   fetchAllBreeds,
   fetchCatImages,
   fetchTotalImageCount,
-} from '../../api/catApi';
-import type { BreedResponse } from '../../api/catApi';
+} from '../../../api/catApi';
+import type { BreedResponse } from '../../../api/catApi';
 
-import { useAppSelector, useAppDispatch } from '../../hooks/reduxHooks';
-import { clearSelection, toggleSelection } from '../../features/selectionSlice';
+import { useAppSelector, useAppDispatch } from '../../../hooks/reduxHooks';
+import { clearSelection, toggleSelection } from '../../../features/selectionSlice';
 
 interface CatCard {
   id: string;
