+++ src/app/components/MainBlock/__mocks__/mocks.ts
