+++ src/app/components/Search/Search.module.scss
