+++ src/app/components/SelectionBlock/SelectionBlock.module.scss
