+++ src/app/components/SelectionBlock/SelectionBlock.tsx
