+++ src/app/components/Spinner/Spinner.module.scss
