+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import ErrorBoundary from '../components/ErrorBoundary/ErrorBoundary';
-import MainBlock from '../components/MainBlock/MainBlock';
+import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
+import MainBlock from './components/MainBlock/MainBlock';
 
 export default function Home() {
   return (
