--- src/components/Pagination/Pagination.tsx
@@ -1,40 +1,0 @@
-import React from 'react';
-import s from './Pagination.module.scss';
-
-interface PaginationProps {
-  currentPage: number;
-  totalPages: number;
-  onPageChange: (page: number) => void;
-}
-
-const Pagination: React.FC<PaginationProps> = ({
-  currentPage,
-  totalPages,
-  onPageChange,
-}) => {
-  return (
-    <div className={s.pagination}>
-      <button
-        onClick={() => onPageChange(currentPage - 1)}
-        disabled={currentPage <= 1}
-        className={s.button}
-      >
-        ← Previous
-      </button>
-
-      <span className={s.page}>
-        {currentPage} из {totalPages}
-      </span>
-
-      <button
-        onClick={() => onPageChange(currentPage + 1)}
-        disabled={currentPage >= totalPages}
-        className={s.button}
-      >
-        Next →
-      </button>
-    </div>
-  );
-};
-
-export default Pagination;
